refactor: extract core types to types.go (#47)
